classification/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aom=3.6.0=h6a678d5_0
appdirs=1.4.4=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blosc=1.21.3=h6a678d5_0
bokeh=3.1.1=py38h2f386ee_0
bottleneck=1.3.7=py38ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py38h27cfd23_1003
brunsli=0.1=h2531618_0
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.3.11=h06a4308_0
cchardet=2.1.7=pypi_0
certifi=2023.5.7=pypi_0
cffi=1.15.1=py38h5eee18b_3
cfitsio=3.470=h5893167_7
chardet=5.1.0=pypi_0
charls=2.2.0=h2531618_0
charset-normalizer=3.1.0=pypi_0
cityscapesscripts=2.2.2=pypi_0
click=8.1.7=py38h06a4308_0
cloudpickle=2.2.1=py38h06a4308_0
colorcet=3.1.0=py38h06a4308_0
coloredlogs=15.0.1=pypi_0
contourpy=1.0.5=py38hdb19cb5_0
cryptography=39.0.1=py38h9ce1e76_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.35=pypi_0
cytoolz=0.12.2=py38h5eee18b_0
dask-core=2023.4.1=py38h06a4308_0
datashader=0.15.2=py38h06a4308_0
datashape=0.5.4=py38h06a4308_1
dav1d=1.2.1=h5eee18b_0
dbus=1.13.18=hb2f20db_0
expat=2.4.9=h6a678d5_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.3.1=py38h06a4308_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.8.0=pypi_0
holoviews=1.17.1=py38h06a4308_0
humanfriendly=10.0=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py38h06a4308_0
imagecodecs=2023.1.23=py38hc4b7b5f_0
imageio=2.31.3=pypi_0
importlib-metadata=8.0.0=pypi_0
importlib_metadata=7.0.1=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.4=py38h06a4308_0
joblib=1.1.1=py38h06a4308_0
jpeg=9e=h5eee18b_1
jsonpatch=1.32=pypi_0
jsonpointer=2.3=pypi_0
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.19.4=h568e23c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libavif=0.11.1=h5eee18b_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcurl=8.1.1=h91b91d3_0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.52.0=ha637b67_1
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libssh2=1.10.0=h37d81fd_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.0=h6a678d5_2
libuuid=1.41.5=h5eee18b_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.3=hcbfbd50_0
libxslt=1.1.37=h2085143_0
libzopfli=1.0.3=he6710b0_0
linkify-it-py=2.0.0=py38h06a4308_0
llvmlite=0.41.1=pypi_0
lmdb=1.5.1=pypi_0
locket=1.0.0=py38h06a4308_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=py38h06a4308_0
markdown-it-py=2.2.0=py38h06a4308_1
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.7.1=py38h06a4308_1
matplotlib-base=3.7.1=py38h417a72b_1
mdit-py-plugins=0.3.0=py38h06a4308_0
mdurl=0.1.0=py38h06a4308_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.6=py38h417a72b_1
mkl_random=1.2.2=py38h417a72b_1
multipledispatch=0.6.0=py38_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
networkx=3.1=py38h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numba=0.58.1=py38h6a678d5_0
numexpr=2.8.4=py38hc78ab66_1
numpngw=0.1.2=pypi_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
opencv-python=4.1.2.30=pypi_0
openjpeg=2.4.0=h9ca470c_1
openssl=1.1.1w=h7f8727e_0
packaging=23.0=py38h06a4308_0
pandas=2.0.1=pypi_0
panel=1.2.3=py38h06a4308_0
param=1.13.0=py38h06a4308_0
partd=1.4.1=py38h06a4308_0
pcre=8.45=h295c915_0
pillow=9.4.0=py38h6a678d5_0
pip=23.0.1=py38h06a4308_0
ply=3.11=py38_0
pooch=1.4.0=pyhd3eb1b0_0
pycocotools=2.0.6=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyct=0.5.0=py38h06a4308_0
pynndescent=0.5.13=pypi_0
pyopenssl=23.0.0=py38h06a4308_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.7=py38h6a678d5_1
pyqt5-sip=12.11.0=py38h6a678d5_1
pyquaternion=0.9.9=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.16=h7a1cb2a_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=1.7.0=py3.8_cuda11.0.221_cudnn8.0.3_0
pytz=2023.3=pypi_0
pyviz_comms=3.0.2=py38h06a4308_0
pywavelets=1.4.1=py38h5eee18b_0
pyyaml=6.0.1=py38h5eee18b_0
qt-main=5.15.2=h8373d8f_8
qt-webengine=5.15.9=hbbf29b9_6
qtwebkit=5.212=h3fafdc1_5
readline=8.2=h5eee18b_0
requests=2.30.0=pypi_0
scikit-image=0.19.3=py38h6a678d5_1
scikit-learn=1.2.2=py38h6a678d5_0
scipy=1.10.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=66.0.0=py38h06a4308_0
sip=6.6.2=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=h6a678d5_1
sqlite=3.41.2=h5eee18b_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2023.4.12=py38h06a4308_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py38h06a4308_0
torchaudio=0.7.0=py38
torchsampler=0.1.2=pypi_0
torchvision=0.8.0=py38_cu110
tornado=6.3.1=pypi_0
tqdm=4.65.0=py38hb070fc8_0
typing=3.7.4.3=pypi_0
typing_extensions=4.5.0=py38h06a4308_0
uc-micro-py=1.0.1=py38h06a4308_0
umap-learn=0.5.6=pypi_0
urllib3=2.0.2=pypi_0
visdom=0.2.4=pypi_0
webencodings=0.5.1=py38_1
websocket-client=1.5.1=pypi_0
wheel=0.38.4=py38h06a4308_0
xarray=2022.11.0=py38h06a4308_0
xyzservices=2022.9.0=py38h06a4308_1
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zfp=1.0.0=h6a678d5_0
zipp=3.11.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
